#77 Sequence refactoring
