+++ Source/Cloud9/Tools/Macro/Operator.h
@@ -1,0 +1,73 @@
+// Copyright (c) 2023 Alexei Gladkikh
+//
+// Permission is hereby granted, free of charge, to any person
+// obtaining a copy of this software and associated documentation
+// files (the "Software"), to deal in the Software without
+// restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following
+// conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+// OTHER DEALINGS IN THE SOFTWARE.
+
+#pragma once
+
+#define OPERATOR_BODY(InOperatorType) \
+	\
+	using OperatorType = InOperatorType; \
+	\
+	template <typename SelfType> \
+	constexpr friend auto operator|(SelfType* Self, InOperatorType&& Operator)\
+	{\
+		return Operator(Self);\
+	}\
+	\
+	template <typename SelfType>\
+	constexpr friend auto operator|(const SelfType* Self, InOperatorType&& Operator)\
+	{\
+		return Operator(Self);\
+	}\
+	\
+	template <typename SelfType>\
+	constexpr friend auto operator|(SelfType& Self, InOperatorType&& Operator)\
+	{\
+		return Operator(Forward<SelfType>(Self));\
+	}\
+	\
+	template <typename SelfType>\
+	constexpr friend auto operator|(SelfType&& Self, InOperatorType&& Operator)\
+	{\
+		return Operator(Forward<SelfType>(Self));\
+	}
+
+namespace Private_TIsOperator
+{
+	/**
+	 * Implementation of std::void_t (not found in UE4)
+	 */
+	template <typename... T>
+	using TVoid = void;
+}
+
+template <typename, typename = void>
+struct TIsOperator
+{
+	enum { Value = false };
+};
+
+template <typename T>
+struct TIsOperator<T, Private_TIsOperator::TVoid<typename T::OperatorType>>
+{
+	enum { Value = true };
+};
